Redid teleop test for new experiment on Monday
What I tested today was if the opmode ran with the old code, and it did, but it seemed stuck in the init loop and never showed telemetry from the init loop: only did the joystick actions. Now, it's separated into init and run loops, so hopefully it should work better. That's what needs to be tested on Monday.